src/main/java/com/bookstore/business/service/impl/AdminDetailsServiceImpl.java
package com.bookstore.business.service.impl;

import com.bookstore.business.repository.AdminRepository;
import com.bookstore.business.repository.model.Admin;
import com.bookstore.security.AdminDetails;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class AdminDetailsServiceImpl implements UserDetailsService {

    private final AdminRepository adminRepository;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        Admin admin = adminRepository.findByUsername(username);
        if (admin == null) {
            throw new UsernameNotFoundException("Admin not found");
        }

        return new AdminDetails(admin);
    }
}
